requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
confection==0.0.4
cymem==2.0.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
frozenlist==1.3.3
h11==0.14.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
Levenshtein==0.20.9
MarkupSafe==2.1.2
multidict==6.0.4
murmurhash==1.0.9
nameparser==1.1.2
nltk==3.8.1
numpy==1.24.2
openai==0.27.2
outcome==1.2.0
packaging==23.0
pathy==0.10.1
preshed==3.0.8
pydantic==1.10.7
pyenchant==3.2.2
PySocks==1.7.1
python-dotenv==1.0.0
rapidfuzz==2.15.0
regex==2023.3.23
requests==2.28.2
selenium==4.8.3
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
textblob==0.17.1
thinc==8.1.9
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.14
wasabi==1.1.1
webdriver-manager==3.8.5
wsproto==1.2.0
yarl==1.8.2
